+++ items/EngineWatchData.h
@@ -6,14 +6,14 @@ using namespace std;
 class EngineWatchDataArchetypeCallback {
     public:
 
-    double index = 0;
+    int index = 0;
     double order = 0;
 
     EngineWatchDataArchetypeCallback(){}
-    EngineWatchDataArchetypeCallback(double index, double order): index(index), order(order){};
+    EngineWatchDataArchetypeCallback(int index, double order): index(index), order(order){};
     EngineWatchDataArchetypeCallback(Json::Value arr){
         if (arr.isNull() == true) return;
-        index = arr["index"].asInt(), order = arr["order"].asInt();
+        index = arr["index"].asInt(), order = arr["order"].asDouble();
     }
 
     Json::Value toJsonObject() {
